fix: Prefer more specific routes in `usePathname` when detecting the currently active pathname for localized pathnames
