+++ packages/next-intl/test/shared/utils.test.tsx
@@ -3,7 +3,8 @@ import {
   hasPathnamePrefixed,
   unprefixPathname,
   matchesPathname,
-  prefixPathname
+  prefixPathname,
+  getSortedPathnames
 } from '../../src/shared/utils';
 
 describe('prefixPathname', () => {
@@ -112,5 +113,63 @@ describe('matchesPathname', () => {
         '/users/23/posts/42/comments'
       )
     ).toBe(false);
+  });
+});
+
+describe('getSortedPathnames', () => {
+  it('works for static routes that include the root', () => {
+    expect(getSortedPathnames(['/', '/foo', '/test'])).toEqual([
+      '/',
+      '/foo',
+      '/test'
+    ]);
+  });
+
+  it('should prioritize non-catch-all routes over catch-all routes', () => {
+    expect(
+      getSortedPathnames(['/categories/[...slug]', '/categories/new'])
+    ).toEqual(['/categories/new', '/categories/[...slug]']);
+  });
+
+  it('should prioritize static routes over optional catch-all routes', () => {
+    expect(
+      getSortedPathnames(['/categories/[[...slug]]', '/categories'])
+    ).toEqual(['/categories', '/categories/[[...slug]]']);
+  });
+
+  it('should prioritize more specific routes over dynamic routes', () => {
+    expect(
+      getSortedPathnames(['/categories/[slug]', '/categories/new'])
+    ).toEqual(['/categories/new', '/categories/[slug]']);
+  });
+
+  it('should prioritize dynamic routes over catch-all routes', () => {
+    expect(
+      getSortedPathnames(['/categories/[...slug]', '/categories/[slug]'])
+    ).toEqual(['/categories/[slug]', '/categories/[...slug]']);
+  });
+
+  it('should prioritize more specific nested routes over dynamic routes', () => {
+    expect(
+      getSortedPathnames([
+        '/articles/[category]/[articleSlug]',
+        '/articles/[category]/new'
+      ])
+    ).toEqual([
+      '/articles/[category]/new',
+      '/articles/[category]/[articleSlug]'
+    ]);
+  });
+
+  it('should prioritize more specific nested routes over catch-all routes', () => {
+    expect(
+      getSortedPathnames([
+        '/articles/[category]/[...articleSlug]',
+        '/articles/[category]/new'
+      ])
+    ).toEqual([
+      '/articles/[category]/new',
+      '/articles/[category]/[...articleSlug]'
+    ]);
   });
 });
